Made stuff go into logical folders Moved source into organized-ish folders and re #included stuff. Fixes issue #17
