+++ Engine/Engine/Engine.vcxproj
@@ -11,42 +11,40 @@
     </ProjectConfiguration>
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="Downcast.cpp" />
+    <ClInclude Include="Defines.hpp" />
+    <ClInclude Include="Game.hpp" />
+    <ClInclude Include="Layers\Layer.hpp" />
+    <ClInclude Include="Layers\LayerManager.hpp" />
+    <ClInclude Include="NameSearchable\NameSearchable.hpp" />
+    <ClInclude Include="Objects\Object.hpp" />
+    <ClInclude Include="Objects\ObjectGroup.hpp" />
+    <ClInclude Include="Objects\ObjectManager.hpp" />
+    <ClInclude Include="ResourceManager\ResourceGroup.hpp" />
+    <ClInclude Include="ResourceManager\ResourceManager.hpp" />
+    <ClInclude Include="TestObject.hpp" />
+    <ClInclude Include="Utility\Downcast.hpp" />
+    <ClInclude Include="Utility\INIParser.hpp" />
+    <ClInclude Include="Utility\Logger.hpp" />
+    <ClInclude Include="Utility\UnitTester.hpp" />
+    <ClInclude Include="Utility\Utilities.hpp" />
+  </ItemGroup>
+  <ItemGroup>
     <ClCompile Include="Game.cpp" />
-    <ClCompile Include="INIParser.cpp" />
     <ClCompile Include="Layers\Layer.cpp" />
     <ClCompile Include="Layers\LayerManager.cpp" />
-    <ClCompile Include="Logger.cpp" />
     <ClCompile Include="Main.cpp" />
     <ClCompile Include="NameSearchable\NameSearchable.cpp" />
-    <ClCompile Include="Object.cpp" />
-    <ClCompile Include="ObjectGroup.cpp" />
-    <ClCompile Include="ObjectManager.cpp" />
+    <ClCompile Include="Objects\Object.cpp" />
+    <ClCompile Include="Objects\ObjectGroup.cpp" />
+    <ClCompile Include="Objects\ObjectManager.cpp" />
     <ClCompile Include="ResourceManager\ResourceGroup.cpp" />
     <ClCompile Include="ResourceManager\ResourceManager.cpp" />
     <ClCompile Include="TestObject.cpp" />
-    <ClCompile Include="UnitTester.cpp" />
-    <ClCompile Include="Utilities.cpp" />
-  </ItemGroup>
-  <ItemGroup>
-    <ClInclude Include="Defines.hpp" />
-    <ClInclude Include="Downcast.hpp" />
-    <ClInclude Include="Game.hpp" />
-    <ClInclude Include="INIParser.hpp" />
-    <ClInclude Include="Layers\Layer.hpp" />
-    <ClInclude Include="Layers\LayerManager.hpp" />
-    <ClInclude Include="Logger.hpp" />
-    <ClInclude Include="NameSearchable\NameSearchable.hpp" />
-    <ClInclude Include="Object.hpp" />
-    <ClInclude Include="ObjectGroup.hpp" />
-    <ClInclude Include="ObjectManager.hpp" />
-    <ClInclude Include="ResourceManager\ResourceGroup.hpp" />
-    <ClInclude Include="ResourceManager\ResourceManager.hpp" />
-    <ClInclude Include="UnitTester.hpp" />
-    <ClInclude Include="Utilities.hpp" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="TestObject.hpp" />
+    <ClCompile Include="Utility\Downcast.cpp" />
+    <ClCompile Include="Utility\INIParser.cpp" />
+    <ClCompile Include="Utility\Logger.cpp" />
+    <ClCompile Include="Utility\UnitTester.cpp" />
+    <ClCompile Include="Utility\Utilities.cpp" />
   </ItemGroup>
   <PropertyGroup Label="Globals">
     <ProjectGuid>{E4773390-6C05-47C2-AEAE-C779CFC9B10F}</ProjectGuid>
