+++ Engine/Engine/Utility/Downcast.hpp
@@ -4,7 +4,7 @@
 #include "boost\shared_ptr.hpp"
 
 //inherited object includes
-#include "TestObject.hpp"
+#include "../TestObject.hpp"
 
 //^^^ those should be combined into group include files ie "Objects.hpp" or this will get annoying
